Define ObjectPoolAllocator in its own file
